chore: move create-validator.ts
